+++ apps/ced-portal-be/CHANGELOG.md
@@ -1,5 +1,23 @@
 # ced-portal-be
 
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies [10cfe73]
+  - @pagopa/io-core-adapter-redis@0.0.3
+
+## 0.0.5
+
+### Patch Changes
+
+- 7187aee: patches
+- Updated dependencies [7187aee]
+  - @pagopa/io-core-adapter-drizzle@0.0.2
+  - @pagopa/io-core-adapter-fastify@0.0.3
+  - @pagopa/io-core-adapter-redis@0.0.2
+  - @pagopa/io-core-domain@0.0.3
+
 ## 0.0.4
 
 ### Patch Changes